requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sentence-transformers==2.2.2
pandas==2.0.3
pymysql==1.1.0
numpy==1.24.3
cachetools==5.3.1
python-json-logger==2.0.7
cryptography==41.0.4
retry==0.9.2
ratelimit==2.2.1
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.2
tqdm==4.66.1
symspellpy==6.7.0
fuzzywuzzy==0.18.0  
requests